moved predicates to response, added File and JSON, minor internal restructurement
